Seperating across components.
